next refactoring step: CampaignManagerInterface + CharityProductManagerInterface = CharityProductManagerInterface unify abstract classes, too Remove all CampaignManager stuff
